Improve diagnostics on lock timeout failure to distinguish between the cases where current process holds the lock and another process holds the lock.
This can help with troubleshooting timeouts in the Gradle test suite. When a test suite uses a mix of embedded execution and tooling API embedded execution in the same test process this can unintentionally cause to the test process attempting to lock a file while already holding the lock. While this isn't intended to happen, it is relatively easy to accidentally cause it to. This is much less of a problem in end user builds.